Use ExoPlayer to display a fake stream
This change adds ExoPlayer as a dependency and
adds a player view to the main activity. This
player view is then used to display a fake
stream. When the fake stream starts, a pre-roll
advertisement is displayed. When this pre-roll
advertisement is completed, the fake stream
resumes. If this adverisement is not correctly
completed, or if the advertisement fails, the
stream is cancelled.