Properly handle events
We were not properly handling the
adFreePod, adError, noAds, and
adCompleted events. We should
only be canceling the stream if
adCompleted is called before
credit is received. If credit
is received prior to adCompleted
being called, we should instead
resume the stream.

In an actual integration, we may
want display the linear ads,
instead of canceling the stream,
if  adFreePod is not called prior
to adCompleted being called.